Moved files from test_foo to eq_ocr, AWS works now
Fixed problems with the fakes3 service not working. It is now a single
line to change to switch to AWS online instead of offline.

Created a new project to bring together all separate parts that we have
been working on. This project includes the c++ random number generator to
start off the functionality for returning a response to the app. The app
now sends and receives a request and response. The request includes the
bitmap data and stroke data for the current strokes on the canvas. The
server side receives this data, saves it to the tmp dir on the server and
then uploads it to the AWS S3 storage. Currently we are working with the
fakes3 server so everything is local and is saved to the buckets folder in
the lib directory.