Initial commit of the current Equation OCR App
This includes an app that was compiled with the FlashDevelop program.
The app currently handles a canvas that records the user's stroke data and
path traveled. Upon clicking the send button the bitmap and stroke
metadata are sent to the Ruby/Sinatra server for server-side processing
and handling (saving to S3 Storage and EB databases).
At the bottom the response and body are returned to the app.